# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chamber Commissions API — سامانه کمیسیون‌های تخصصی
  version: 1.0.0
  description: 'REST API of the commissions platform. All business endpoints require `Authorization: Bearer <accessToken>`.


    Real-time: Socket.IO at `/socket.io` (auth: `{ token }`). Emit `meeting:join` with a meeting id, then receive `meeting.updated`, `attendance.updated`
    (officers), `quorum.updated`, `agenda.updated`, `comment.created`, `vote.opened`, `vote.progress`, `vote.closed`, and personal `notification`
    events.


    Errors: `{ error: { code, message, details? } }` with Persian messages.'
servers:
- url: http://localhost:4000
security:
- bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: &id001
              type: string
            message: *id001
            details:
              type: array
              items:
                type: object
                properties:
                  path: *id001
                  message: *id001
...
paths:
...
  /api/me/home:
    get:
      tags:
      - Dashboard
      summary: خانه من (موبایل)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications:
    get:
      tags:
      - Notifications
      summary: اعلان‌ها (?unread)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications/{id}/read:
    post:
      tags:
      - Notifications
      summary: خوانده شد
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
  /api/notifications/read-all:
    post:
      tags:
      - Notifications
      summary: خواندن همه
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/dashboard/commission:
    get:
      tags:
      - Dashboard
      summary: داشبورد کمیسیون (?commissionId)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/dashboard/chamber:
    get:
      tags:
      - Dashboard
      summary: داشبورد اتاق (?chamberId)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reports/resolutions:
    get:
      tags:
      - Reports
      summary: گزارش مصوبات (?format=csv)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reports/attendance:
    get:
      tags:
      - Reports
      summary: گزارش حضور اعضا (?commissionId)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/search:
    get:
      tags:
      - Reports
      summary: جستجوی سراسری (?q)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/audit:
    get:
      tags:
      - Audit
      summary: رویدادنگاری (?chamberId&entity&entityId&action)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/audit/verify:
    get:
      tags:
      - Audit
      summary: بررسی سلامت زنجیره هش
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'